requirements.txt
arrow==1.2.2
black==22.1.0
certifi==2021.10.8        
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
greenlet==1.1.2
idna==3.3
mypy-extensions==0.4.3    
numpy==1.22.3
pandas==1.4.1
pathspec==0.9.0
platformdirs==2.5.1       
psycopg2==2.9.3
python-dateutil==2.8.2    
python-decouple==3.6
python-dotenv==0.20.0
pytz==2021.3
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.32
tomli==2.0.1
typing_extensions==4.1.1
urllib3==1.26.8